Python/django_nuxt_test/backend/requirements.txt
django==3.2
djangorestframework
django-cors-headers
psycopg2>=2.8